+++ packages/expo-modules-core/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - Expose the embedded fingerprint to native code through `EmbeddedFingerprint`. ([#49494](https://github.com/expo/expo/pull/49494) by [@vonovak](https://github.com/vonovak))
+- [iOS] Add `FingerprintCheckProtocol`, the literals the dev-launcher responder and the expo-linking filter both match on. ([#49494](https://github.com/expo/expo/pull/49494) by [@vonovak](https://github.com/vonovak))
 
 ### 🐛 Bug fixes